Use existing definitions instead of using the manifest
